% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcc_show_bicluster_hscore.R
\name{funcc_show_bicluster_hscore}
\alias{funcc_show_bicluster_hscore}
\title{plotting hscore of each bi-cluster on bicluster dimension}
\usage{
funcc_show_bicluster_hscore(fun_mat, res_input)
}
\arguments{
\item{fun_mat}{The data array (n x m x T) where each entry corresponds to the measure of one observation i, i=1,...,n, for a functional variable m, m=1,...,p, at point t, t=1,...,T}

\item{res_input}{An object produced by the funcc_biclust function}
}
\value{
a figure representing the dimensions of each bi-cluster (i.e. number of rows and columns)
}
\description{
funcc_show_bicluster_hscore graphically shows the hscore vs the dimension (i.e. number of rows and columns) of each bi-cluster
}
\examples{
 
data("funCCdata")
res <- funcc_biclust(funCCdata,delta=10,theta=1,alpha=1,beta=0,const_alpha=TRUE)
funcc_show_bicluster_hscore(funCCdata,res)

}
