% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{funCCdata}
\alias{funCCdata}
\title{Simulated data}
\format{An object of class \code{array} of dimension 30 x 7 x 240.}
\usage{
data(funCCdata)
}
\description{
funCC.data is a functional dataset displaying block structure
}
\examples{
data(funCCdata)

}
\keyword{datasets}
